requirements.txt
pony>=0.6.1
pyTelegramBotAPI>=0.3.5
PyYaml>=3.11